extract server mod
