+++ Cargo.toml
@@ -31,3 +31,8 @@ serde_json = "1.0"
 
 [dev-dependencies]
 httpmock = "0.6"
+criterion = "0.3"
+
+[[bench]]
+name = "my_benchmark"
+harness = false
