+++ src/parser.rs
@@ -1,4 +1,4 @@
-use crate::Frame;
+use crate::frame::Frame;
 
 use std::{str, vec};
