+++ tests/integration_test.rs
@@ -1,0 +1,4 @@
+#[test]
+fn it_adds_two() {
+    assert_eq!(4, 4);
+}
